+++ app/backend/lib/emails/templates/assesmentSecondReviewChange.ts
@@ -6,7 +6,7 @@ import {
 const assesmentSecondReviewChange: EmailTemplateProvider = (
   applicationId: string,
   url: string,
-  initiator: string,
+  initiator: any,
   params: any
 ): EmailTemplate => {
   const { ccbcNumber } = params;
@@ -14,10 +14,10 @@ const assesmentSecondReviewChange: EmailTemplateProvider = (
     emailTo: [34, 71], // Temporary IDs to handle email recipients
     emailCC: [],
     tag: 'assesment-second-review-change',
-    subject: `${initiator} requested a 2nd Review for Eligibility Screening - ${ccbcNumber}`,
+    subject: `${initiator?.givenName} requested a 2nd Review for Eligibility Screening - ${ccbcNumber}`,
     body: `
-    <h1>${initiator} requested a 2nd Review for Eligibility Screening - ${ccbcNumber}</h1>
-    <p>${initiator} requested a 2nd Review for Eligibility Screening - ${ccbcNumber}, <a href='${url}/analyst/application/${applicationId}/assessments/screening'>Click here</a> to view the Eligibility Screening in the CCBC Portal<p>
+    <h1>${initiator?.givenName} requested a 2nd Review for Eligibility Screening - ${ccbcNumber}</h1>
+    <p>${initiator?.givenName} requested a 2nd Review for Eligibility Screening - ${ccbcNumber}, <a href='${url}/analyst/application/${applicationId}/assessments/screening'>Click here</a> to view the Eligibility Screening in the CCBC Portal<p>
   `,
   };
 };