+++ app/tests/pages/applicantportal/form/[page].test.tsx
@@ -213,12 +213,15 @@ describe('The form page', () => {
             ccbcNumber: 'CCBC-010001',
             projectName: 'Project testing title',
             updatedAt: '2022-08-15T16:43:28.973734-04:00',
+            intakeByIntakeId: {
+              ccbcIntakeNumber: 3,
+              closeTimestamp: '2022-09-06T23:59:59-07:00',
+            },
             formData: {
               formByFormSchemaId: {
                 jsonSchema: schema,
               },
             },
-            intakeByIntakeId: null,
           },
           allForms: {
             nodes: [
@@ -230,7 +233,7 @@ describe('The form page', () => {
           },
           openIntake: {
             closeTimestamp: '2022-08-27T12:52:00.00000-04:00',
-            ccbcIntakeNumber: 2,
+            ccbcIntakeNumber: 3,
           },
           session: {
             sub: '4e0ac88c-bf05-49ac-948f-7fd53c7a9fd6',
@@ -247,7 +250,7 @@ describe('The form page', () => {
     pageTestingHelper.loadQuery(payload);
     pageTestingHelper.renderPage();
 
-    const projectAreasText = screen.getByText(/within zones 6,7,8/);
+    const projectAreasText = screen.getByText(/within zones 9,10/);
     expect(projectAreasText).toBeInTheDocument();
   });
 
@@ -305,6 +308,62 @@ describe('The form page', () => {
     expect(projectAreasText).toBeInTheDocument();
   });
 
+  it('shows intake 4 project area text', async () => {
+    const payload = {
+      Query() {
+        return {
+          applicationByRowId: {
+            status: 'draft',
+            ccbcNumber: 'CCBC-010001',
+            projectName: 'Project testing title',
+            updatedAt: '2022-08-15T16:43:28.973734-04:00',
+            formData: {
+              formByFormSchemaId: {
+                jsonSchema: schema,
+              },
+            },
+            intakeByIntakeId: null,
+          },
+          allForms: {
+            nodes: [
+              {
+                rowId: 10,
+                jsonSchema: schemaV2,
+              },
+            ],
+          },
+          openIntake: null,
+          allIntakes: {
+            edges: [
+              {
+                node: {
+                  ccbcIntakeNumber: 4,
+                  closeTimestamp: '2022-08-27T12:52:00.00000-04:00',
+                },
+              },
+            ],
+          },
+          session: {
+            sub: '4e0ac88c-bf05-49ac-948f-7fd53c7a9fd6',
+          },
+        };
+      },
+    };
+    pageTestingHelper.setMockRouterValues({
+      query: { id: '1', page: '2' },
+    });
+
+    jest.spyOn(moduleApi, 'useFeature').mockReturnValue(mockAcceptedZones);
+
+    pageTestingHelper.loadQuery(payload);
+    pageTestingHelper.renderPage();
+
+    const projectAreasText = screen.getByText(
+      /CCBC will accept applications for all eligible areas in the Province/
+    );
+    expect(projectAreasText).toBeInTheDocument();
+  });
+
   it('shows accepted project areas and handles modal', async () => {
     pageTestingHelper.setMockRouterValues({
       query: { id: '1', page: '2' },
@@ -344,6 +403,10 @@ describe('The form page', () => {
         return {
           id: 'TestApplicationId',
           status: 'submitted',
+          intakeByIntakeId: {
+            ccbcIntakeNumber: 3,
+            closeTimestamp: '2022-09-06T23:59:59-07:00',
+          },
           formData: {
             formByFormSchemaId: {
               jsonSchema: schema,
@@ -437,6 +500,10 @@ describe('The form page', () => {
         return {
           id: 'TestApplicationId',
           status: 'submitted',
+          intakeByIntakeId: {
+            ccbcIntakeNumber: 3,
+            closeTimestamp: '2022-09-06T23:59:59-07:00',
+          },
           formData: {
             formByFormSchemaId: {
               jsonSchema: schema,
@@ -519,6 +586,10 @@ describe('The form page', () => {
         return {
           id: 'TestApplicationId',
           status: 'submitted',
+          intakeByIntakeId: {
+            ccbcIntakeNumber: 3,
+            closeTimestamp: '2022-09-06T23:59:59-07:00',
+          },
           formData: {
             formByFormSchemaId: {
               jsonSchema: schema,