+++ CHANGELOG.md
@@ -1,3 +1,46 @@
+# [1.154.0](https://github.com/bcgov/CONN-CCBC-portal/compare/v1.153.0...v1.154.0) (2024-04-15)
+
+### Bug Fixes
+
+- datacatalog url ([4a5d846](https://github.com/bcgov/CONN-CCBC-portal/commit/4a5d8465b03e696712948c976690eecb4171a72f))
+- email template user format ([b940e77](https://github.com/bcgov/CONN-CCBC-portal/commit/b940e77dde2ffe6ce5d3f5e3b24c2b1b7ee9a6ca))
+- remove intake 4 zone 14 kmz links ([8d113e8](https://github.com/bcgov/CONN-CCBC-portal/commit/8d113e8ffba541f1348afe9bbce85bd3bb963422))
+- update logic to only display error in related fields ([9ccf251](https://github.com/bcgov/CONN-CCBC-portal/commit/9ccf251e16cc1ae1f9aedf95c2ee1b74f31a8ccd))
+
+### Features
+
+- review shows incorrect value not only blank ([818f908](https://github.com/bcgov/CONN-CCBC-portal/commit/818f90876b18d01b7030b0c0b737981745f7c423))
+- schema changes for funding sources ([19a3ba8](https://github.com/bcgov/CONN-CCBC-portal/commit/19a3ba8254b2c7f19da7720499b09077d2f2cd6d))
+- show error message on funding sources ([a1ae0a0](https://github.com/bcgov/CONN-CCBC-portal/commit/a1ae0a05eb9dcb583cafa5357f0e8e480c8c548a))
+
+# [1.153.0](https://github.com/bcgov/CONN-CCBC-portal/compare/v1.152.0...v1.153.0) (2024-04-11)
+
+### Features
+
+- add padding for intake 4 image ([c1830f9](https://github.com/bcgov/CONN-CCBC-portal/commit/c1830f940e54b65f18fe951a128a7bb69e29f9de))
+- default for intake 4 ([4ac361d](https://github.com/bcgov/CONN-CCBC-portal/commit/4ac361d6ca21e622edb96051755ad3aee58c612e))
+- intake 4 uses new map ([730fb9e](https://github.com/bcgov/CONN-CCBC-portal/commit/730fb9ebd8578d66e51eb2bfe796914fe5b49b1d))
+
+# [1.152.0](https://github.com/bcgov/CONN-CCBC-portal/compare/v1.150.0...v1.152.0) (2024-04-11)
+
+### Bug Fixes
+
+- 2nd review colour not showing on assessment dash ([d924da8](https://github.com/bcgov/CONN-CCBC-portal/commit/d924da888b495c1a65c214466f0be848ffc37b1f))
+- add guard for previously submitted applications from previous intakes ([3ea3099](https://github.com/bcgov/CONN-CCBC-portal/commit/3ea30994a5a63950156a4734250a3ec3f6b57f46))
+- add schema ([4b67ec5](https://github.com/bcgov/CONN-CCBC-portal/commit/4b67ec5547c0adca1564a2dcc02d7db7e4ea396f))
+- better sort function for zones ([50446dc](https://github.com/bcgov/CONN-CCBC-portal/commit/50446dc1a618c7fc784407bac1827516dafcf267))
+- changes to schema ([ac2dd7c](https://github.com/bcgov/CONN-CCBC-portal/commit/ac2dd7c88ce8806ca2233be2635f4a663b34cf9d))
+- failing e2e test ([6066410](https://github.com/bcgov/CONN-CCBC-portal/commit/60664108b5e6b9a1ca6d779ef40bd587596ed93f))
+- project information edit button issue ([cb28444](https://github.com/bcgov/CONN-CCBC-portal/commit/cb284449855fa77d7ee8fe4e4317b9b85deb6df9))
+
+### Features
+
+- add email notification to second review eligibility screening ([ef3389f](https://github.com/bcgov/CONN-CCBC-portal/commit/ef3389f5910192488ce8b5fe9e7f9d88309149f6))
+- add logic so draft and intake 4 intakes are using latest ui schema ([feb6e01](https://github.com/bcgov/CONN-CCBC-portal/commit/feb6e01f7afbbb38d5bf3858bab54ef4aed49e73))
+- add other files upload to funding agreement and change request ([8e6c2d0](https://github.com/bcgov/CONN-CCBC-portal/commit/8e6c2d0c62b830ebe0380ef1e782ebef59ab2c58))
+- modal does not appear for intake 4 ([4d8ca38](https://github.com/bcgov/CONN-CCBC-portal/commit/4d8ca38de45a22b2821754c54b4b51c9a7ae5c3f))
+- new schema for intake ([46bb86d](https://github.com/bcgov/CONN-CCBC-portal/commit/46bb86de0670de82572e12d87a427dd767c2e266))
+
 # [1.151.0](https://github.com/bcgov/CONN-CCBC-portal/compare/v1.150.0...v1.151.0) (2024-04-09)
 
 ### Bug Fixes